requirements.txt
playwright
python-dotenv
pydantic
playwright-stealth>=1.0.0
transformers
Pillow
pytesseract
# agent-workflow-memory>=0.1.0 # This was the old memory system, replaced by mem0ai
mem0ai
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
ollama
llm-reasoners
sentence-transformers
